+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - Added `h5` view
     - Added `beforeTabs` and `afterTabs` config options to `tabs` view
     - Fixed report header jumping on hover in edit mode on report page in some browsers
+    - Added virtual view `render` to render a view with config as current data and data as previous data
 
 ## 1.0.0-beta.13 (18-06-2019)